Move payment functions to db file
